requirements.txt
aiohttp==3.8.3
aiosignal==1.2.0
alembic==1.8.1
anyio==3.6.1
async-timeout==4.0.2
asyncpg==0.26.0
attrs==22.1.0
base58==2.1.1
bitarray==2.6.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
comtypes==1.1.14
cryptography==38.0.1
cytoolz==0.12.0
databases==0.6.1
ecdsa==0.18.0
eth-abi==2.2.0
eth-account==0.5.9
eth-hash==0.5.0
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.9.5
fastapi==0.85.0
fastapi-jwt-auth==0.5.0
frozenlist==1.3.1
greenlet==1.1.3
h11==0.13.0
hexbytes==0.3.0
idna==3.4
ipfshttpclient==0.8.0a2
jsonschema==4.16.0
lru-dict==1.1.8
Mako==1.2.3
MarkupSafe==2.1.1
multiaddr==0.0.9
multidict==6.0.2
netaddr==0.8.0
ormar==0.11.3
parsimonious==0.8.1
protobuf==3.19.5
psycopg2-binary==2.9.3
pyasn1==0.4.8
pycparser==2.21
pycrypto==2.6.1
pycryptodome==3.15.0
pydantic==1.10.2
PyJWT==1.7.1
pyrsistent==0.18.1
python-jose==3.3.0
requests==2.28.1
rlp==2.0.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.41
starlette==0.20.4
toolz==0.12.0
typing_extensions==4.3.0
urllib3==1.26.12
uvicorn==0.18.3
varint==1.0.2
web3==5.31.1
websockets==9.1
yarl==1.8.1
